pyproject.toml
[project]
name = "ble-locator-server"
version = "0.1.0"
description = "Headless BLE locator server with MQTT ingestion, configurable algorithms, and CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=1.6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.10.3",
    "pyinstaller>=6.14.2",
    "pyyaml>=6.0.2",
    "scipy>=1.16.2",
]

[project.scripts]
ble-locator-server = "ble_locator_server.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ble_locator_server*"]
